requirements.txt
coverage==4.4.1
Django==1.11.3
django-bootstrap3==9.0.0
django-crispy-forms==1.6.1
pytz==2017.2
